Externalise testTask scripts and config
If they are inlined in the pipeline file, we cannot be sure which
version of the script/configs have run in which test run.

+++ entrypoint.sh
@@ -285,7 +285,12 @@ kind::gen_config() {
 
   if [ -n "$userConf" ]
   then
-    echo "$userConf" > "$kindConfFile"
+    if [ -e "$userConf" ]
+    then
+      cat "$userConf" > "$kindConfFile"
+    else
+      echo "$userConf" > "$kindConfFile"
+    fi
   else
     cat "$defaultConfFile" > "$kindConfFile"
   fi
